+++ sdk/resourcemanager/cdn/armcdn/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armcdn
 
 const (
 	moduleName    = "armcdn"
-	moduleVersion = "v1.0.0"
+	moduleVersion = "v2.0.0"
 )
 
 // AFDEndpointProtocols - Supported protocols for the customer's endpoint.